+++ scripts/wfc_analysis_2026/combine_uplift_per_steer.py
@@ -5,7 +5,7 @@ import numpy as np
 import pandas as pd
 from scipy.stats import norm
 
-from scripts.wake_steering_analysis.hot_wake_steering_helpers import _calc_yaw_stats
+from scripts.wfc_analysis_2026.hot_wake_steering_helpers import _calc_yaw_stats
 
 logger = logging.getLogger(__name__)
