+++ scripts/wfc_analysis_2026/floris_config/hot_emgauss.yaml
