+++ scripts/wfc_analysis_2026/floris_config/turbine_inputs/SGRE-2.3-82.yaml
