+++ scripts/wfc_analysis_2026/inspect_data.py
