+++ scripts/wfc_analysis_2026/wind_up_config/HOT_dynamic_yaw.yaml
