+++ scripts/wfc_analysis_2026/wind_up_config/asset/HOT.yaml
