+++ scripts/wfc_analysis_2026/wind_up_config/asset/masts_and_lidars/ZTM_5060.yaml
