+++ scripts/wfc_analysis_2026/wind_up_config/asset/masts_and_lidars/ZX300_2428.yaml
