+++ scripts/wfc_analysis_2026/wind_up_config/asset/turbine_type/SWT_2p3_82.yaml
