+++ scripts/wfc_analysis_2026/wind_up_config/exclusion_periods/HOT_DY.yaml
